pyproject.toml
[tool.poetry]
name = "pyre"
version = "0.1.0"
description = "FIRE with python"
authors = ["dithyrambe <dithyrambe@outlook.fr>"]
readme = "README.md"
packages = [
  {include = "pyre", from = "src"}
]

[tool.poetry.scripts]
pyre = "pyre.cli.main:app"

[tool.poetry.dependencies]
python = ">=3.11,<4"
numpy = "^1.26.4"
pandas = "^2.2.2"
typer = "^0.12.3"
nicegui = {extras = ["highcharts"], version = "^1.4.24"}
pendulum = "^3.0.0"
yfinance = "^0.2.40"
plotille = "^5.0.0"
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
pyyaml = "^6.0.1"
pydantic-settings = "^2.3.1"
pydantic = "^2.7.3"
uvicorn = "^0.30.1"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
ipython = "^8.22.1"
pytest = "^8.1.1"
testcontainers = "^4.5.1"
ruff = "^0.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry-dynamic-versioning]
enable = true
pattern  = "^(?P<base>\\d+\\.\\d+\\.\\d+)(-?((?P<stage>[a-zA-Z]+)\\.?(?P<revision>\\d+)?))?"

[tool.ruff]
line-length = 100

[tool.pyright]
venvPath = "."
venv = ".venv"
